add username and name to sign up form for ios
