--- apple/OmnivoreKit/Sources/App/Views/Registration/EmailAuthView.swift
@@ -1,263 +1,0 @@
-import Models
-import Services
-import SwiftUI
-import Utils
-import Views
-
-enum EmailAuthState {
-  case signIn
-  case signUp
-  case loading
-}
-
-@MainActor final class EmailAuthViewModel: ObservableObject {
-  @Published var loginError: LoginError?
-  @Published var emailAuthState = EmailAuthState.loading
-
-  func loadAuthState() {
-    // check tokens here to determine pending/active/no user
-    emailAuthState = .signUp
-  }
-
-  func submitCredentials(
-    email: String,
-    password: String,
-    authenticator: Authenticator
-  ) async {
-    do {
-      try await authenticator.submitEmailLogin(email: email, password: password)
-    } catch {
-      loginError = error as? LoginError
-    }
-  }
-}
-
-struct EmailAuthView: View {
-  @Environment(\.presentationMode) private var presentationMode
-  @StateObject private var viewModel = EmailAuthViewModel()
-
-  @ViewBuilder var primaryContent: some View {
-    switch viewModel.emailAuthState {
-    case .signUp:
-      EmailSignupFormView(viewModel: viewModel)
-    case .signIn:
-      EmailLoginFormView(viewModel: viewModel)
-    case .loading:
-      VStack {
-        Spacer()
-        ProgressView()
-        Spacer()
-      }
-    }
-  }
-
-  var body: some View {
-    NavigationView {
-      ZStack {
-        Color.appBackground.edgesIgnoringSafeArea(.all)
-        primaryContent
-          .frame(maxWidth: 300)
-          .navigationBarTitleDisplayMode(.inline)
-          .toolbar {
-            ToolbarItem(placement: .barTrailing) {
-              Button(
-                action: { presentationMode.wrappedValue.dismiss() },
-                label: { Image(systemName: "xmark").foregroundColor(.appGrayTextContrast) }
-              )
-            }
-          }
-      }
-    }
-    .task {
-      viewModel.loadAuthState()
-    }
-  }
-}
-
-struct EmailLoginFormView: View {
-  enum FocusedField {
-    case email, password
-  }
-
-  @Environment(\.horizontalSizeClass) var horizontalSizeClass
-  @EnvironmentObject var authenticator: Authenticator
-  @ObservedObject var viewModel: EmailAuthViewModel
-
-  @FocusState private var focusedField: FocusedField?
-  @State private var email = ""
-  @State private var password = ""
-
-  var body: some View {
-    VStack(spacing: 0) {
-      VStack(spacing: 28) {
-        ScrollView(showsIndicators: false) {
-          if horizontalSizeClass == .regular {
-            Spacer(minLength: 150)
-          }
-          VStack {
-            VStack(alignment: .leading, spacing: 6) {
-              Text("Email")
-                .font(.appFootnote)
-                .foregroundColor(.appGrayText)
-              TextField("", text: $email)
-                .textContentType(.emailAddress)
-                .textInputAutocapitalization(.never)
-                .disableAutocorrection(true)
-                .focused($focusedField, equals: .email)
-            }
-            .padding(.bottom, 8)
-
-            VStack(alignment: .leading, spacing: 6) {
-              Text("Password")
-                .font(.appFootnote)
-                .foregroundColor(.appGrayText)
-              SecureField("", text: $password)
-                .textContentType(.password)
-                .textInputAutocapitalization(.never)
-                .disableAutocorrection(true)
-                .focused($focusedField, equals: .password)
-            }
-            .padding(.bottom, 16)
-
-            Button(
-              action: {
-                Task {
-                  await viewModel.submitCredentials(
-                    email: email,
-                    password: password,
-                    authenticator: authenticator
-                  )
-                }
-              },
-              label: { Text("Submit") }
-            )
-            .buttonStyle(SolidCapsuleButtonStyle(color: .appDeepBackground, width: 300))
-
-            if let loginError = viewModel.loginError {
-              LoginErrorMessageView(loginError: loginError)
-            }
-
-            HStack {
-              Button(
-                action: { viewModel.emailAuthState = .signUp },
-                label: {
-                  Text("Don't have an account?")
-                    .foregroundColor(.appGrayTextContrast)
-                    .underline()
-                }
-              )
-              .padding(.vertical)
-              Spacer()
-            }
-          }
-          .textFieldStyle(StandardTextFieldStyle())
-          .onSubmit {
-            if focusedField == .email {
-              focusedField = .password
-            } else {
-              focusedField = nil
-            }
-          }
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity)
-
-        Spacer()
-      }
-    }
-    .navigationTitle("Sign In")
-  }
-}
-
-struct EmailSignupFormView: View {
-  enum FocusedField {
-    case email, password, fullName, username
-  }
-
-  @Environment(\.horizontalSizeClass) var horizontalSizeClass
-  @EnvironmentObject var authenticator: Authenticator
-  @ObservedObject var viewModel: EmailAuthViewModel
-
-  @FocusState private var focusedField: FocusedField?
-  @State private var email = ""
-  @State private var password = ""
-
-  var body: some View {
-    VStack(spacing: 0) {
-      VStack(spacing: 28) {
-        ScrollView(showsIndicators: false) {
-          if horizontalSizeClass == .regular {
-            Spacer(minLength: 150)
-          }
-          VStack {
-            VStack(alignment: .leading, spacing: 6) {
-              Text("Email")
-                .font(.appFootnote)
-                .foregroundColor(.appGrayText)
-              TextField("", text: $email)
-                .textContentType(.emailAddress)
-                .textInputAutocapitalization(.never)
-                .disableAutocorrection(true)
-                .focused($focusedField, equals: .email)
-            }
-            .padding(.bottom, 8)
-
-            VStack(alignment: .leading, spacing: 6) {
-              Text("Password")
-                .font(.appFootnote)
-                .foregroundColor(.appGrayText)
-              SecureField("", text: $password)
-                .textContentType(.password)
-                .textInputAutocapitalization(.never)
-                .disableAutocorrection(true)
-                .focused($focusedField, equals: .password)
-            }
-            .padding(.bottom, 16)
-
-            Button(
-              action: {
-                Task {
-                  await viewModel.submitCredentials(
-                    email: email,
-                    password: password,
-                    authenticator: authenticator
-                  )
-                }
-              },
-              label: { Text("Submit") }
-            )
-            .buttonStyle(SolidCapsuleButtonStyle(color: .appDeepBackground, width: 300))
-
-            if let loginError = viewModel.loginError {
-              LoginErrorMessageView(loginError: loginError)
-            }
-
-            HStack {
-              Button(
-                action: { viewModel.emailAuthState = .signIn },
-                label: {
-                  Text("Already have an account?")
-                    .foregroundColor(.appGrayTextContrast)
-                    .underline()
-                }
-              )
-              .padding(.vertical)
-              Spacer()
-            }
-          }
-          .textFieldStyle(StandardTextFieldStyle())
-          .onSubmit {
-            if focusedField == .email {
-              focusedField = .password
-            } else {
-              focusedField = nil
-            }
-          }
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity)
-
-        Spacer()
-      }
-    }
-    .navigationTitle("Sign Up")
-  }
-}
